pacman/requirements.txt
backports.functools-lru-cache==1.5
bottle==0.12.13
certifi==2018.4.16
chardet==3.0.4
click==5.1
colorama==0.3.9
cycler==0.10.0
drawnow==0.72.0
idna==2.6
kiwisolver==1.0.1
lockfile==0.12.2
matplotlib==2.2.2
numpy==1.14.2
platformio==3.5.2
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
semantic-version==2.6.0
six==1.11.0
subprocess32==3.2.7
urllib3==1.22
